pyproject.toml
[project]
name = "fusio-worker-runtime"
version = "0.2.1"
authors = [
  { name="Christoph Kappestein", email="christoph.kappestein@gmail.com" },
]
description = "Fusio worker Python runtime"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Code Generators",
]
dependencies = [
    "httpx >= 0.27",
    "pydantic >= 2.7.0",
    "mysql-connector-python >= 9.2.0",
    "psycopg >= 3.0",
    "pymongo >= 4.8",
    "elasticsearch >= 8.14"
]

[project.urls]
Homepage = "https://github.com/apioo/fusio-worker-python-runtime"
Issues = "https://github.com/apioo/fusio-worker-python-runtime/issues"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["runtime", "runtime.generated"]
package-dir = {"" = "src"}
